+++ app/models/user.rb
@@ -20,6 +20,8 @@ class User < ActiveRecord::Base
   has_many :tokens, :class_name=>"OauthToken", :order=>"authorized_at desc", :include=>[:client_application]
   # => oauth support
 
+  record_activities
+
   protected
 
     def create_profile