+++ test/unit/user_test.rb
@@ -1,0 +1,7 @@
+require File.dirname(__FILE__) + '/../test_helper'
+
+class UserTest < Test::Unit::TestCase
+  context "A User" do
+    should_have_many :activities
+  end
+end