+++ app/controllers/profiles_controller.rb
@@ -1,25 +1,26 @@
 class ProfilesController < ApplicationController
-  
+
   def index
     @order = params[:order] || 'profiles.created_at'
     @page = params[:page] || '1'
     @asc = params[:asc] || 'desc'   
     @profiles = Profile.active.paginate :per_page => Tog::Config["plugins.tog_social.profile.list.page.size"],
                                  :page => @page,
-                                 :order => @order + " " + @asc 
+                                 :order => "profiles.#{@order} #{@asc}"
+
     respond_to do |format|
       format.html # index.html.erb
       format.xml  { render :xml => @profiles }
-    end    
+    end
   end
-  
+
   def show
-    @profile = Profile.find(params[:id])
+    @profile = Profile.active.find(params[:id])
     store_location
     respond_to do |format|
       format.html # index.html.erb
       format.xml  { render :xml => @profile }
-    end    
+    end
   end
-  
+
 end