+++ app/helpers/groups_helper.rb
@@ -12,8 +12,8 @@ module GroupsHelper
 
   def image_for_group(group, size, options={})
     return if !group
-    if group.image
-      photo_url = url_for_image_column(group, "image", :name => size)
+    if group.image?
+      photo_url = group.image.url(size)
       options.merge!(:alt => "Photo for group: #{group.name}")
       return image_tag(photo_url, options) if photo_url
     else