+++ app/models/group.rb
@@ -24,14 +24,14 @@ class Group < ActiveRecord::Base
 
   before_create :set_default_image
 
-  file_column :image, :root_path => File.join(RAILS_ROOT, "public/system/group_photos"), :web_root => 'system/group_photos/', :magick => {
-      :versions => {
-        :big    => {:crop => "1:1", :size => Tog::Config["plugins.tog_social.group.image.versions.big"],    :name => "big"},
-        :medium => {:crop => "1:1", :size => Tog::Config["plugins.tog_social.group.image.versions.medium"], :name => "medium"},
-        :small  => {:crop => "1:1", :size => Tog::Config["plugins.tog_social.group.image.versions.small"],  :name => "small" },
-        :tiny   => {:crop => "1:1", :size => Tog::Config["plugins.tog_social.group.image.versions.tiny"],   :name => "tiny"  }
-      }
-  }
+  has_attached_file :image, {
+    :url => "/system/:class/:attachment/:id/:style_:basename.:extension",
+    :styles => { 
+      :big    => Tog::Plugins.settings(:tog_social, "group.image.versions.big"),
+      :medium => Tog::Plugins.settings(:tog_social, "group.image.versions.medium"),
+      :small  => Tog::Plugins.settings(:tog_social, "group.image.versions.small"),
+      :tiny   => Tog::Plugins.settings(:tog_social, "group.image.versions.tiny")
+    }}.merge(Tog::Plugins.storage_options)
 
   record_activity_of :user
   acts_as_abusable
@@ -117,7 +117,7 @@ class Group < ActiveRecord::Base
   end
 
   def set_default_image
-    unless self.image
+    unless self.image?
       if Tog::Config["plugins.tog_social.group.image.default"]
         default_group_image = File.join(RAILS_ROOT, 'public', 'tog_social', 'images', Tog::Config["plugins.tog_social.group.image.default"])
         self.image = File.new(default_group_image)