+++ config/routes.rb
@@ -1,4 +1,4 @@
-# Add your custom routes here.  If in config/routes.rb you would 
+# Add your custom routes here.  If in config/routes.rb you would
 # add <tt>map.resources</tt>, here you would add just <tt>resources</tt>
 
 resources :profiles
@@ -29,3 +29,13 @@ end
 namespace(:admin) do |admin|
   admin.resources :groups, :member => { :activate => :post}
 end
+
+# => oauth support
+resources :oauth_clients
+authorize '/oauth/authorize',:controller=>'oauth',:action=>'authorize'
+request_token '/oauth/request_token',:controller=>'oauth',:action=>'request_token'
+access_token '/oauth/access_token',:controller=>'oauth',:action=>'access_token'
+test_request '/oauth/test_request',:controller=>'oauth',:action=>'test_request'
+x1 '/oauth_test', :controller=>'oauth_test',:action=>'test'
+x2 '/oauth_callback', :controller=>'oauth_test',:action=>'callback'
+# => oauth support