+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Edge
 ----
+* has_many :activities added to user thanks to record\_activities macro provided by acts\_as\_scribe. kudos to John Paul for catch this one.
 
 0.4.0
 ----