+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Edge
 ----
+* Fixed #117. Full name and user's login used on moderated group joined notification.
 * has_many :activities added to user thanks to record\_activities macro provided by acts\_as\_scribe. kudos to John Paul for catch this one.
 
 0.4.0